move history utilities
